Fixed CAN-FD, added even more verbose logging, added ability to put strings into preferences, added ability to zap preferences for only one device, a lot of work to implement comm to the ESP32

+++ src/devices/misc/SystemDevice.h
@@ -41,7 +41,7 @@ public:
     uint16_t adcOffset[NUM_ANALOG];
     uint32_t canSpeed[4];
     uint8_t swcanMode; //should can0 be in SWCAN mode?
-    uint8_t logLevel;
+    int16_t logLevel;
 };
 
 class SystemDevice: public Device {
